Add files via upload
I added a few functions to help with the logic of the game.I  also added a new variable in the state class to represent what are points in hearts.  These are the new functions that I added to the class. 
Update_ScoreBoard -- use after the trick to update everyone's points
Trick_lead - find the person who won the trick and adjust the state
Legal moves - The cards that a player can play given a suit(which i called domain)