Leading is kinda sorta done.
The agent will attempt to lead with its lowest card out of the suit with the lowest number of cards. If Diamonds and Clubs are both void, it will try to lead with Hearts, then Spades.

For following, I was thinking that we may want to have this agent follow with its high cards (if it's safe to do so) in order to get rid of problem cards if it can earlier on